cv_model/requirements.txt
torch>=2.0.0
ultralytics>=8.0.0
roboflow>=1.1.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

